Use Webpack to build website assets
