+++ .codeclimate.yml
@@ -1,5 +1,10 @@
 version: "2"
 plugins:
+  eslint:
+    enabled: true
+    channel: eslint-6
+    config:
+      config: natlas-server/.eslintrc.yml
   pep8:
     enabled: true
     checks:
@@ -13,7 +18,7 @@ plugins:
         enabled: false
     exclude_patterns:
     # machine generated, duplication likely
-    - "natlas-server/migrations/versions/" 
+    - "natlas-server/migrations/versions/"
   radon:
     enabled: true
     config:
@@ -32,4 +37,4 @@ plugins:
           python_version: 3
     exclude_patterns:
     # machine generated, duplication likely
-    - "natlas-server/migrations/versions/" 
+    - "natlas-server/migrations/versions/"
