+++ .gitignore
@@ -18,3 +18,4 @@ natlas-server/media
 natlas-server/logs
 natlas-agent/logs
 tmp
+natlas-server/node_modules
