+++ natlas-server/.eslintrc.yml
@@ -1,0 +1,20 @@
+---
+env:
+  es6: true
+  browser: true
+parserOptions:
+  ecmaVersion: 2018
+  sourceType: module
+settings:
+  import/resolver: webpack
+plugins:
+  - import
+extends:
+  - eslint:recommended
+  - plugin:import/errors
+  - plugin:import/warnings
+  - plugin:import/typescript
+rules:
+  indent:
+  - error
+  - tab
