+++ natlas-server/app/static/natlas.scss
@@ -1,8 +1,11 @@
 /*
 * Natlas custom styles
-* Really needs to be organized and minified
+* Really needs to be organized
 *
 */
+@import '~@fortawesome/fontawesome-free/css/fontawesome.css';
+@import '~@fortawesome/fontawesome-free/css/solid.css';
+@import '~bootstrap/dist/css/bootstrap.css';
 
 .natlas-pink {
   color:#ed1e79;
