+++ natlas-server/package.json
@@ -1,0 +1,26 @@
+{
+	"dependencies": {
+		"@fortawesome/fontawesome-free": "^5.13.0",
+		"bootstrap": "^4.4.1",
+		"datatables.net": "^1.10.20",
+		"jquery": "^3.4.1",
+		"popper.js": "^1.16.0"
+	},
+	"devDependencies": {
+		"@typescript-eslint/parser": "^2.29.0",
+		"css-loader": "^3.5.2",
+		"eslint": "^6.8.0",
+		"eslint-import-resolver-webpack": "^0.12.1",
+		"eslint-plugin-import": "^2.20.2",
+		"file-loader": "^6.0.0",
+		"mini-css-extract-plugin": "^0.9.0",
+		"node-sass": "^4.14.0",
+		"sass-loader": "^8.0.2",
+		"ts-loader": "^7.0.1",
+		"typescript": "^3.8.3",
+		"webpack": "4.43.0",
+		"webpack-cli": "^3.3.11",
+		"webpack-dev-server": "^3.10.3",
+		"webpack-yam-plugin": "^1.0.1"
+	}
+}
